+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `useDSHeading` hook to render a default H2 heading or a custom heading element.
 - Adds the `sizeBasedOn` prop to the `Image` component.
 - Adds the `isDarkBackgroundImage` prop to the `Hero` component.
+- Adds optional `placeholder` and `placeholderTo` props to the `DatePicker` component.
 
 ### Updates
 
@@ -29,6 +30,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `FeaturedContent` component by adjusting the spacing in the `"fullScreen"` variant to better align the component text content with the page text content.
 - Updates the `"campaign"` variant of the `Hero` component to improve the spacing around the component.
 - Updates the `Card` component so that it accepts the `imageProps.isLazy` prop and passes it to its internal `Image` component.
+- Updates the `DatePicker`'s `initialDate` and `initialDateTo` props to accept an empty string.
+
 
 ## Fixes