+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
-- Fixes bug where month and year `Datepicker` calendars were being displayed vertically rather than horizontally.
+- Fixes bug where month and year `Datepicker` calendars were rendering vertically rather than horizontally.
 
 ## 2.1.3 (December 7, 2023)